댓글

+++ hello-spring/src/main/java/com/hello/forum/bbs/web/BoardController.java
@@ -1,27 +1,50 @@
 package com.hello.forum.bbs.web;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.List;
+
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.xssf.streaming.SXSSFWorkbook;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.hello.forum.bbs.service.BoardService;
 import com.hello.forum.bbs.vo.BoardListVO;
 import com.hello.forum.bbs.vo.BoardVO;
 import com.hello.forum.beans.FileHandler;
+import com.hello.forum.exceptions.MakeXlsxFileException;
+import com.hello.forum.exceptions.PageNotFoundException;
+import com.hello.forum.member.vo.MemberVO;
+import com.hello.forum.utils.AjaxResponse;
 import com.hello.forum.utils.ValidationUtils;
+
+import io.github.seccoding.excel.option.WriteOption;
+import io.github.seccoding.excel.write.ExcelWrite;
 
 //import jakarta.validation.Valid;
 
 @Controller
 public class BoardController {
+	
+	private Logger logger = LoggerFactory.getLogger(BoardController.class);
 	
 	@Autowired
 	private FileHandler fileHandler;
@@ -88,9 +111,9 @@ public class BoardController {
 			/* @Valid에 의해 실행된 파라미터 검사(NotEmpty, Email, Size, Min, Max 등)의 결과 */
 //			BindingResult bindingResult,
 			@RequestParam MultipartFile file,
-			Model model
-			) {
-		System.out.println("글 등록 처리를 해야합니다.");
+			@SessionAttribute("_LOGIN_USER_") MemberVO memberVO,
+			Model model) {
+		logger.info("글 등록 처리를 해야합니다.");
 		/* Servlet Like
 		 * HttpServletRequest를 이용
 		 * - Interceptor에서 이용
@@ -115,9 +138,7 @@ public class BoardController {
 		// 수동 검사 시작
 		// 제목 검사
 		boolean isNotEmptySubject = ValidationUtils.notEmpty(boardVO.getSubject());
-		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
 		boolean isNotEmptyContent = ValidationUtils.notEmpty(boardVO.getContent());
-		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
 		
 		if (!isNotEmptySubject) {
 			// 제목을 입력하지 않았다면
@@ -125,14 +146,7 @@ public class BoardController {
 			model.addAttribute("boardVO", boardVO);
 			return "board/boardwrite";
 		}
-		
-		if (!isNotEmptyEmail) {
-			// 이메일을 입력하지 않았다면
-			model.addAttribute("errorMessage", "이메일은 필수 입력 값입니다.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardwrite";
-		}
-		
+				
 		if (!isNotEmptyContent) {
 			// 내용을 입력하지 않았다면
 			model.addAttribute("errorMessage", "내용은 필수 입력 값입니다.");
@@ -140,19 +154,14 @@ public class BoardController {
 			return "board/boardwrite";
 		}
 		
-		if (!isEmailFormat) {
-			// 이메일을 이메일 형태로 입력하지 않았다면
-			model.addAttribute("errorMessage", "이메일을 올바른 형태로 작성해주세요.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardwrite";
-		}
+		boardVO.setEmail(memberVO.getEmail());
 		
 		boolean isCreateSuccess = this.boardService.createNewBoard(boardVO, file);
 		if (isCreateSuccess) {
-			System.out.println("글 등록 성공!");
+			logger.info("글 등록 성공!");
 		}
 		else {
-			System.out.println("글 등록 실패!");
+			logger.info("글 등록 실패!");
 		}
 		
 		// board/boardlist 페이지를 보여주는 URL로 이동처리
@@ -185,9 +194,14 @@ public class BoardController {
 	}
 	
 	@GetMapping("/board/modify/{id}")	// /board/modify/1	<-- id 변수의 값은 1
-	public String viewBoardModifyPage(@PathVariable int id, Model model) {
+	public String viewBoardModifyPage(@PathVariable int id, Model model, 
+									  @SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
 		// 1. 전달받은 id의 값으로 게시글을 조회한다.
 		BoardVO boardVO = this.boardService.getOneBoard(id, false);
+		
+		if (!memberVO.getEmail().equals(boardVO.getEmail())) {
+			throw new PageNotFoundException();
+		}
 		
 		// 2. 게시글의 정보를 화면에 보내준다.
 		model.addAttribute("boardVO", boardVO);
@@ -203,25 +217,23 @@ public class BoardController {
 	 * @return
 	 */
 	@PostMapping("/board/modify/{id}")
-	public String doBoardModify(@PathVariable int id, BoardVO boardVO, @RequestParam MultipartFile file, Model model) {
+	public String doBoardModify(@PathVariable int id, BoardVO boardVO, 
+								@RequestParam MultipartFile file, Model model, 
+								@SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
+		
+		BoardVO originalBoardVO = this.boardService.getOneBoard(id, false);
+		if (!originalBoardVO.getEmail().equals(memberVO.getEmail())) {
+			throw new PageNotFoundException();
+		}
 		
 		// 수동 검사 시작
 		// 제목 검사
 		boolean isNotEmptySubject = ValidationUtils.notEmpty(boardVO.getSubject());
-		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
 		boolean isNotEmptyContent = ValidationUtils.notEmpty(boardVO.getContent());
-		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
 		
 		if (!isNotEmptySubject) {
 			// 제목을 입력하지 않았다면
 			model.addAttribute("errorMessage", "제목은 필수 입력 값입니다.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardmodify";
-		}
-		
-		if (!isNotEmptyEmail) {
-			// 이메일을 입력하지 않았다면
-			model.addAttribute("errorMessage", "이메일은 필수 입력 값입니다.");
 			model.addAttribute("boardVO", boardVO);
 			return "board/boardmodify";
 		}
@@ -233,24 +245,16 @@ public class BoardController {
 			return "board/boardmodify";
 		}
 		
-		if (!isEmailFormat) {
-			// 이메일을 이메일 형태로 입력하지 않았다면
-			model.addAttribute("errorMessage", "이메일을 올바른 형태로 작성해주세요.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardmodify";
-		}
-		
 		// Command Object에는 전달된 ID가 없으므로
 		// PathVariable로 전달된 ID를 셋팅해준다.
-		boardVO.setId(id);
-		
+		boardVO.setId(id);		
 		boolean isUpdatedSuccess = this.boardService.updateOneBoard(boardVO, file);
 		
 		if (isUpdatedSuccess) {
-			System.out.println("수정 성공했습니다!");
+			logger.info("수정 성공했습니다!");
 		}
 		else {
-			System.out.println("수정 실패했습니다!");
+			logger.info("수정 실패했습니다!");
 		}
 		
 		return "redirect:/board/view?id=" + id;
@@ -271,15 +275,21 @@ public class BoardController {
 	 * URL이나 링크 등으로 데이터를 조회하거나 삭제할 경우 -> GET
 	 */
 	@GetMapping("/board/delete/{id}")
-	public String doDeleteboard(@PathVariable int id) {
+	public String doDeleteboard(@PathVariable int id, 
+								@SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
+		
+		BoardVO originalBoardVO = this.boardService.getOneBoard(id, false);
+		if (!originalBoardVO.getEmail().equals(memberVO.getEmail())) {
+			throw new PageNotFoundException();
+		}
 		
 		boolean isDeletedSuccess = this.boardService.deleteOneBoard(id);
 		
 		if (isDeletedSuccess) {
-			System.out.println("게시글 삭제 성공.");
+			logger.info("게시글 삭제 성공.");
 		}
 		else {
-			System.out.println("게시글 삭제 실패.");
+			logger.info("게시글 삭제 실패.");
 		}
 		
 		return "redirect:/board/list";
@@ -293,16 +303,125 @@ public class BoardController {
 		
 		// 만약, 게시글이 존재하지 않다면 "잘못된 접근입니다."라는 에러를 사용자에게 보여준다.
 		if (boardVO == null) {
-			throw new IllegalArgumentException("잘못된 접근입니다.");
+			throw new PageNotFoundException();
 		}
 		
 		// 첨부된 파일이 없을 경우에도 "잘못된 접근입니다."라는 에러를 사용자에게 보여준다.
 		if (boardVO.getFileName() == null || boardVO.getFileName().length() == 0) {
-			throw new IllegalArgumentException("잘못된 접근입니다.");
+			throw new PageNotFoundException();
 		}
 		
 		// 첨부된 파일이 있을 경우엔 파일을 사용자에게 보내준다. (Download)
 		return this.fileHandler.download(boardVO.getOriginFileName(), boardVO.getFileName());
+	}
+	
+	@GetMapping("/board/excel/download2")
+	public ResponseEntity<Resource> downloadExcelFile2() {
+		BoardListVO boardListVO = boardService.getAllBoard();
+		
+		WriteOption<BoardVO> writeOption = new WriteOption<>();
+		writeOption.setFileName("게시글_목록.xlsx");
+		writeOption.setFilePath("C:\\uploadFiles");
+		writeOption.setContents(boardListVO.getBoardList());
+		
+		File excelFile = ExcelWrite.write(writeOption);
+		
+		return this.fileHandler.download("게시글_목록.xlsx", excelFile.getName());
+	}
+	
+	@GetMapping("/board/excel/download")
+	public ResponseEntity<Resource> downloadExcelFile() {
+		
+		// 모든 게시글 조회
+		BoardListVO boardListVO = boardService.getAllBoard();
+		
+		// XLSX 문서 만들기
+		Workbook workbook = new SXSSFWorkbook(-1);
+		
+		Sheet sheet = workbook.createSheet("게시글 목록");
+		
+		Row row = sheet.createRow(0);
+		
+		Cell cell = row.createCell(0);
+		cell.setCellValue("번호");
+		
+		cell = row.createCell(1);
+		cell.setCellValue("제목");
+		
+		cell = row.createCell(2);
+		cell.setCellValue("첨부파일명");
+		
+		cell = row.createCell(3);
+		cell.setCellValue("작성자이메일");
+		
+		cell = row.createCell(4);
+		cell.setCellValue("조회수");
+		
+		cell = row.createCell(5);
+		cell.setCellValue("등록일");
+		
+		cell = row.createCell(6);
+		cell.setCellValue("수정일");
+		
+		List<BoardVO> boardList = boardListVO.getBoardList();
+		int rowIndex = 1;
+		for (BoardVO boardVO : boardList) {
+			row = sheet.createRow(rowIndex);
+			cell = row.createCell(0);
+			cell.setCellValue("" + boardVO.getId());
+			
+			cell = row.createCell(1);
+			cell.setCellValue("" + boardVO.getSubject());
+			
+			cell = row.createCell(2);
+			cell.setCellValue("" + boardVO.getOriginFileName());
+			
+			cell = row.createCell(3);
+			cell.setCellValue("" + boardVO.getEmail());
+			
+			cell = row.createCell(4);
+			cell.setCellValue("" + boardVO.getViewCnt());
+			
+			cell = row.createCell(5);
+			cell.setCellValue("" + boardVO.getCrtDt());
+			
+			cell = row.createCell(6);
+			cell.setCellValue("" + boardVO.getMdfyDt());
+			
+			rowIndex += 1;
+		}
+		
+		File storedFile = fileHandler.getStoredFile("게시글_목록.xlsx");
+		OutputStream os = null;
+		try {
+			os = new FileOutputStream(storedFile);
+			workbook.write(os);
+		} catch (IOException e) {
+			throw new MakeXlsxFileException();
+		} finally {
+			try {
+				workbook.close();
+			} catch (IOException e) {}
+			if (os != null) {
+				try {
+					os.flush();
+				} catch (IOException e) {}
+				try {
+					os.close();
+				} catch (IOException e) {}
+			}
+		}
+		
+		return this.fileHandler.download("게시글_목록.xlsx", "게시글_목록.xlsx");
+	}
+	
+	@ResponseBody
+	@PostMapping("/ajax/board/excel/write")
+	public AjaxResponse doExcelUpload(@RequestParam MultipartFile excelFile) {
+
+		boolean isSuccess = this.boardService.createMassiveBoard2(excelFile);
+
+		return new AjaxResponse().append("result", isSuccess).append("next", "/board/list");
 	}
 
 }
